--isolatedModulesエラーを回避するため、空のindex.tsxをなくす
