+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import "./App.css";
 import { AppProvider } from "@/providers";
-import { AppRoutes } from "@/routes/public";
+import { AppRoutes } from "@/routes";
 
 function App() {
   const [count, setCount] = useState(0);
