--- src/routes/public.tsx
@@ -1,16 +1,0 @@
-import { useRoutes } from "react-router-dom";
-
-import { ConceptRoutes } from "@/features/concept/routes";
-
-const routes = [
-  {
-    path: "/",
-    element: <ConceptRoutes />,
-  },
-];
-
-export const AppRoutes = () => {
-  const element = useRoutes([...routes]);
-
-  return <>{element}</>;
-};
